+++ src/main/java/no/nav/familie/integrasjoner/kodeverk/KodeverkController.kt
@@ -1,4 +1,4 @@
-package no.nav.familie.ks.oppslag.kodeverk
+package no.nav.familie.integrasjoner.kodeverk
 
 import no.nav.security.token.support.core.api.ProtectedWithClaims
 import org.springframework.http.MediaType
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.*
 
 @RestController
 @ProtectedWithClaims(issuer = "azuread")
-@RequestMapping(path = ["/api/kodeverk/poststed"], produces = [MediaType.APPLICATION_JSON_VALUE])
+@RequestMapping(path = ["/api/kodeverk/poststed"], produces = [MediaType.TEXT_PLAIN_VALUE])
 class PostnummerController(val kodeverkService: KodeverkService) {
 
     @GetMapping("/{postnummer}")
