+++ src/main/java/no/nav/familie/integrasjoner/kodeverk/domene/PostnummerDto.kt
@@ -1,4 +1,4 @@
-package no.nav.familie.ks.oppslag.kodeverk.domene
+package no.nav.familie.integrasjoner.kodeverk.domene
 
 import com.fasterxml.jackson.annotation.JsonProperty
